+++ src/main/java/hillel/lesson10/PokerOOP.java
@@ -1,0 +1,15 @@
+package hillel.lesson10;
+
+public class PokerOOP {
+
+    public static void main(String[] args) {
+        try {
+            Deck deck = new Deck();
+            deck.createDeck();
+            deck.shuffleDeck();
+            deck.serveCards();
+        } catch (Exceptions.EmptyDeckException | Exceptions.InvalidCardIndexException e) {
+            System.out.println(e.getMessage());
+        }
+    }
+}
