+++ src/main/java/hillel/lesson2/CircleArea.java
@@ -1,0 +1,10 @@
+package hillel.lesson2;
+
+public class CircleArea {
+
+    public static void main(String[] args) {
+        double radius = 9;
+        double area = Math.PI * Math.pow(radius,2);
+        System.out.println("Площв круга дорівнює " + area);
+    }
+}
