+++ src/main/java/hillel/lesson3/CheckOddEven.java
@@ -1,0 +1,11 @@
+package hillel.lesson3;
+
+public class CheckOddEven {
+
+    public static void main(String[] args) {
+        int number = 10;
+        String result = number %2 == 0? "Even number": "Odd number";
+        System.out.println(result + "\nBYE!");
+
+    }
+}
