+++ src/main/java/hillel/lesson4/SinX.java
@@ -1,0 +1,9 @@
+package hillel.lesson4;
+
+public class SinX {
+    public static void main(String[] args) {
+        for (int i = 0; i < 361; i = i+10) {
+            System.out.println("sin"+ " " +i);
+        }
+    }
+}
