+++ src/main/java/hillel/lesson6/PokerOOP.java
@@ -1,0 +1,14 @@
+package hillel.lesson6;
+
+
+public class PokerOOP {
+
+            public static void main(String[] args) {
+            Deck deck = new Deck();
+            deck.createDeck();
+            deck.shuffleDeck();
+            deck.serveCards();
+        }
+
+
+}
