+++ src/test/java/tests/TestClass.java
@@ -1,0 +1,12 @@
+package tests;
+
+import org.testng.annotations.Test;
+
+public class TestClass {
+
+    @Test
+    public void testClass () {
+
+    System.out.println("My first test!");
+    }
+}
