--- src/ua/hillel/lesson10/Card.java
@@ -1,22 +1,0 @@
-package ua.hillel.lesson10;
-class Card {
-    private final String suit;
-    private final String value;
-
-    public Card(String suit, String value) {
-        this.suit = suit;
-        this.value = value;
-    }
-
-    public String getSuit() {
-        return suit;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public String toString() {
-        return value + " of " + suit;
-    }
-}
