--- src/ua/hillel/lesson2/TriangleArea.java
@@ -1,11 +1,0 @@
-package ua.hillel.lesson2;
-
-public class TriangleArea {
-
-    public static void main(String[] args) {
-
-        int sideA = 5;
-        int sideB = 8;
-        System.out.println("Площа трикутника " + sideB*sideA/2);
-    }
-}
