--- src/ua/hillel/lesson3/PrintMonthInWordSwitch.java
@@ -1,53 +1,0 @@
-package ua.hillel.lesson3;
-
-public class PrintMonthInWordSwitch {
-
-    public static void main(String[] args) {
-
-        int numberMonth = 11;
-        String month;
-
-        switch (numberMonth) {
-            case 1:
-                month = "January";
-                break;
-            case 2:
-                month = "February";
-                break;
-            case 3:
-                month = "March";
-                break;
-            case 4:
-                month = "April";
-                break;
-            case 5:
-                month = "May";
-                break;
-            case 6:
-                month = "June";
-                break;
-            case 7:
-                month = "July";
-                break;
-            case 8:
-                month = "August";
-                break;
-            case 9:
-                month = "September";
-                break;
-            case 10:
-                month = "October";
-                break;
-            case 11:
-                month = "November";
-                break;
-            case 12:
-                month = "December";
-                break;
-            default:
-                month = "Not a valid month";
-
-        }
-        System.out.println(month);
-    }
-}
