--- src/ua/hillel/lesson4/SinX.java
@@ -1,9 +1,0 @@
-package ua.hillel.lesson4;
-
-public class SinX {
-    public static void main(String[] args) {
-        for (int i = 0; i < 361; i = i+10) {
-            System.out.println("sin"+ " " +i);
-        }
-    }
-}
