--- src/ua/hillel/lesson8/Color.java
@@ -1,11 +1,0 @@
-package ua.hillel.lesson8;
-
-public class Color {
-    public static final String ANSI_RESET = "\u001B[0m";
-    public static final String ANSI_RED = "\u001B[31m";
-    public static final String ANSI_GREEN = "\u001B[32m";
-    public static final String ANSI_YELLOW = "\u001B[33m";
-    public static final String ANSI_BLUE = "\u001B[34m";
-
-}
-
