--- src/ua/hillel/lesson8/Level.java
@@ -1,21 +1,0 @@
-package ua.hillel.lesson8;
-
-public enum Level {
-
-    INFO ("Information: "),
-    DEBUG("Debugging: "),
-    WARN("Warning: "),
-    ERROR("Error: ");
-    private final String message;
-
-    public String getMessage;
-
-    Level(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
-
